+++ R/biodivMapR_tiling.R
@@ -30,12 +30,12 @@
 #' @importFrom progressr with_progress progressor handlers
 #' @export
 
-biodivMapR_full_tiles <- function(feature_dir, list_features, mask_dir = NULL,
-                                  output_dir, window_size, plots, nbCPU = 1,
-                                  site_name = NULL, options = NULL){
+biodivMapR_tiling <- function(feature_dir, list_features, mask_dir = NULL,
+                              output_dir, window_size, plots, nbCPU = 1,
+                              site_name = NULL, options = NULL){
 
   # define options
-  options <- set_options_biodivMapR(fun = 'biodivMapR_full_tiles', options = options)
+  options <- set_options_biodivMapR(fun = 'biodivMapR_tiling', options = options)
   nb_clusters <- options$nb_clusters
   nb_samples_alpha<- options$nb_samples_alpha
   nb_samples_beta <- options$nb_samples_beta
@@ -86,113 +86,179 @@ biodivMapR_full_tiles <- function(feature_dir, list_features, mask_dir = NULL,
     nbCPU <-  maxCPU
 
   if (!is.null(alpha_metrics) | !is.null(beta_metrics)){
-    if (nbCPU>1){
+    if (nbCPU>1) {
       cl <- parallel::makeCluster(nbCPU)
       parallel::clusterEvalQ(cl, {library(biodivMapR)})
       with(plan("cluster", workers = cl), local = TRUE)
-      handlers("cli")
-      with_progress({
-        p <- progressr::progressor(steps = maxCPU)
-        future.apply::future_lapply(X = samples$ID_aoi,
-                                    FUN = run_biodivMapR_plot,
-                                    feature_dir = feature_dir,
-                                    mask_dir = mask_dir,
-                                    list_features = list_features,
-                                    Kmeans_info = Kmeans_info,
-                                    Beta_info = Beta_info,
-                                    alpha_metrics = alpha_metrics,
-                                    Hill_order = Hill_order,
-                                    beta_metrics = beta_metrics,
-                                    fd_metrics = NULL,
-                                    output_dir = output_dir,
-                                    window_size = window_size,
-                                    maxRows = maxRows,
-                                    pcelim = pcelim,
-                                    moving_window = moving_window, p = p,
-                                    future.seed = TRUE,
-                                    future.chunk.size = NULL,
-                                    future.scheduling = structure(TRUE,
-                                                                  ordering = "random"))
-      })
+      funct <- future.apply::future_lapply
+    } else {
+      funct <- lapply
+    }
+    handlers("cli")
+    with_progress({
+      p <- progressr::progressor(steps = maxCPU)
+      funct(X = samples$ID_aoi,
+            FUN = run_biodivMapR_plot,
+            feature_dir = feature_dir,
+            mask_dir = mask_dir,
+            list_features = list_features,
+            Kmeans_info = Kmeans_info,
+            Beta_info = Beta_info,
+            alpha_metrics = alpha_metrics,
+            Hill_order = Hill_order,
+            beta_metrics = beta_metrics,
+            fd_metrics = NULL,
+            output_dir = output_dir,
+            window_size = window_size,
+            maxRows = maxRows,
+            pcelim = pcelim,
+            moving_window = moving_window, p = p,
+            future.seed = TRUE,
+            future.chunk.size = NULL,
+            future.scheduling = 1)
+    })
+    if (nbCPU>1) {
       parallel::stopCluster(cl)
       plan(sequential)
-    } else if (nbCPU==1){
-      handlers("cli")
-      with_progress({
-        p <- progressr::progressor(steps = maxCPU)
-        lapply(X = samples$ID_aoi,
-               FUN = run_biodivMapR_plot,
-               feature_dir = feature_dir,
-               mask_dir = mask_dir,
-               list_features = list_features,
-               Kmeans_info = Kmeans_info,
-               Beta_info = Beta_info,
-               alpha_metrics = alpha_metrics,
-               Hill_order = Hill_order,
-               beta_metrics = beta_metrics,
-               fd_metrics = NULL,
-               output_dir = output_dir,
-               window_size = window_size,
-               maxRows = maxRows,
-               moving_window = moving_window, p = p)
-      })
     }
+
+    #
+    # if (nbCPU>1){
+    #   handlers("cli")
+    #   with_progress({
+    #     p <- progressr::progressor(steps = maxCPU)
+    #     future.apply::future_lapply(X = samples$ID_aoi,
+    #                                 FUN = run_biodivMapR_plot,
+    #                                 feature_dir = feature_dir,
+    #                                 mask_dir = mask_dir,
+    #                                 list_features = list_features,
+    #                                 Kmeans_info = Kmeans_info,
+    #                                 Beta_info = Beta_info,
+    #                                 alpha_metrics = alpha_metrics,
+    #                                 Hill_order = Hill_order,
+    #                                 beta_metrics = beta_metrics,
+    #                                 fd_metrics = NULL,
+    #                                 output_dir = output_dir,
+    #                                 window_size = window_size,
+    #                                 maxRows = maxRows,
+    #                                 pcelim = pcelim,
+    #                                 moving_window = moving_window, p = p,
+    #                                 future.seed = TRUE,
+    #                                 future.chunk.size = NULL,
+    #                                 future.scheduling = 1)
+    #   })
+    #   parallel::stopCluster(cl)
+    #   plan(sequential)
+    # } else if (nbCPU==1){
+    #   handlers("cli")
+    #   with_progress({
+    #     p <- progressr::progressor(steps = maxCPU)
+    #     lapply(X = samples$ID_aoi,
+    #            FUN = run_biodivMapR_plot,
+    #            feature_dir = feature_dir,
+    #            mask_dir = mask_dir,
+    #            list_features = list_features,
+    #            Kmeans_info = Kmeans_info,
+    #            Beta_info = Beta_info,
+    #            alpha_metrics = alpha_metrics,
+    #            Hill_order = Hill_order,
+    #            beta_metrics = beta_metrics,
+    #            fd_metrics = NULL,
+    #            output_dir = output_dir,
+    #            window_size = window_size,
+    #            maxRows = maxRows,
+    #            moving_window = moving_window, p = p)
+    #   })
+    # }
   }
 
   if (!is.null(fd_metrics)){
     for (fd_metric in fd_metrics){
       message(paste('computing functional metric: ', fd_metric))
-      if (nbCPU>1){
+      if (nbCPU>1) {
         cl <- parallel::makeCluster(nbCPU)
         parallel::clusterEvalQ(cl, {library(biodivMapR)})
         with(plan("cluster", workers = cl), local = TRUE)
-        handlers("cli")
-        with_progress({
-          p <- progressr::progressor(steps = maxCPU)
-          future.apply::future_lapply(X = samples$ID_aoi,
-                                      FUN = run_biodivMapR_plot,
-                                      feature_dir = feature_dir,
-                                      mask_dir = mask_dir,
-                                      list_features = list_features,
-                                      Kmeans_info = Kmeans_info,
-                                      Beta_info = NULL,
-                                      alpha_metrics = NULL,
-                                      Hill_order = Hill_order,
-                                      beta_metrics = NULL,
-                                      fd_metrics = fd_metric,
-                                      output_dir = output_dir,
-                                      window_size = window_size,
-                                      maxRows = maxRows,
-                                      pcelim = pcelim,
-                                      moving_window = moving_window, p = p,
-                                      future.seed = TRUE,
-                                      future.chunk.size = NULL,
-                                      future.scheduling = structure(TRUE,
-                                                                    ordering = "random"))
-        })
-        parallel::stopCluster(cl)
-        plan(sequential)
-      } else if (nbCPU==1){
-        handlers("cli")
-        with_progress({
-          p <- progressr::progressor(steps = maxCPU)
-          lapply(X = samples$ID_aoi,
-                 FUN = run_biodivMapR_plot,
-                 feature_dir = feature_dir,
-                 mask_dir = mask_dir,
-                 list_features = list_features,
-                 Kmeans_info = Kmeans_info,
-                 Beta_info = NULL,
-                 alpha_metrics = NULL,
-                 beta_metrics = NULL,
-                 Hill_order = Hill_order,
-                 fd_metrics = fd_metric,
-                 output_dir = output_dir,
-                 window_size = window_size,
-                 maxRows = maxRows,
-                 moving_window = moving_window, p = p)
-        })
+        funct <- future.apply::future_lapply
+      } else {
+        funct <- lapply
       }
+      handlers("cli")
+      with_progress({
+        p <- progressr::progressor(steps = maxCPU)
+        funct(X = samples$ID_aoi,
+              FUN = run_biodivMapR_plot,
+              feature_dir = feature_dir,
+              mask_dir = mask_dir,
+              list_features = list_features,
+              Kmeans_info = Kmeans_info,
+              Beta_info = NULL,
+              alpha_metrics = NULL,
+              Hill_order = Hill_order,
+              beta_metrics = NULL,
+              fd_metrics = fd_metric,
+              output_dir = output_dir,
+              window_size = window_size,
+              maxRows = maxRows,
+              pcelim = pcelim,
+              moving_window = moving_window, p = p,
+              future.seed = TRUE,
+              future.chunk.size = NULL,
+              future.scheduling = 1)
+      })
+
+
+      # if (nbCPU>1){
+      #   cl <- parallel::makeCluster(nbCPU)
+      #   parallel::clusterEvalQ(cl, {library(biodivMapR)})
+      #   with(plan("cluster", workers = cl), local = TRUE)
+      #   handlers("cli")
+      #   with_progress({
+      #     p <- progressr::progressor(steps = maxCPU)
+      #     future.apply::future_lapply(X = samples$ID_aoi,
+      #                                 FUN = run_biodivMapR_plot,
+      #                                 feature_dir = feature_dir,
+      #                                 mask_dir = mask_dir,
+      #                                 list_features = list_features,
+      #                                 Kmeans_info = Kmeans_info,
+      #                                 Beta_info = NULL,
+      #                                 alpha_metrics = NULL,
+      #                                 Hill_order = Hill_order,
+      #                                 beta_metrics = NULL,
+      #                                 fd_metrics = fd_metric,
+      #                                 output_dir = output_dir,
+      #                                 window_size = window_size,
+      #                                 maxRows = maxRows,
+      #                                 pcelim = pcelim,
+      #                                 moving_window = moving_window, p = p,
+      #                                 future.seed = TRUE,
+      #                                 future.chunk.size = NULL,
+      #                                 future.scheduling = structure(TRUE,
+      #                                                               ordering = "random"))
+      #   })
+      #   parallel::stopCluster(cl)
+      #   plan(sequential)
+      # } else if (nbCPU==1){
+      #   handlers("cli")
+      #   with_progress({
+      #     p <- progressr::progressor(steps = maxCPU)
+      #     lapply(X = samples$ID_aoi,
+      #            FUN = run_biodivMapR_plot,
+      #            feature_dir = feature_dir,
+      #            mask_dir = mask_dir,
+      #            list_features = list_features,
+      #            Kmeans_info = Kmeans_info,
+      #            Beta_info = NULL,
+      #            alpha_metrics = NULL,
+      #            beta_metrics = NULL,
+      #            Hill_order = Hill_order,
+      #            fd_metrics = fd_metric,
+      #            output_dir = output_dir,
+      #            window_size = window_size,
+      #            maxRows = maxRows,
+      #            moving_window = moving_window, p = p)
+      #   })
+      # }
     }
   }