+++ R/get_raster_diversity_mw.R
@@ -25,6 +25,12 @@ get_raster_diversity_mw <- function(
     beta_metrics = 'bray', fd_metrics = NULL, window_size, maxRows = NULL,
     pcelim = 0.02, nbCPU = 1, min_sun = 0.25){
 
+  if (!is.null(fd_metrics)){
+    message('functional diversity metrics are not too computationally intensive for moving windows')
+    message('setting fd_metrics = FALSE')
+    fd_metrics <- FALSE
+  }
+
   # prepare to read input raster data
   r_in <- list()
   if (is.null(names(input_raster_path)))
@@ -95,21 +101,6 @@ get_raster_diversity_mw <- function(
       inputdata <- do.call(what = rbind, args = inputdata)
     }
 
-    # inputdata <- list()
-    # for (i in 1:ncol(rast_in)){
-    #   xmin <- max(c(1, i-(window_size-1)/2))
-    #   xmax <- min(c(ncol(rast_in), i+(window_size-1)/2))
-    #   ymin <- max(c(1, j-(window_size-1)/2))
-    #   ymax <- min(c(nrow(rast_in), j+(window_size-1)/2))
-    #   inputdata[[i]] <- na.omit(rast_in[ymin:ymax, xmin:xmax, ])
-    #   if (nrow(inputdata[[i]])>0)
-    #     inputdata[[i]]$win_ID <- i
-    # }
-    # ll <- unlist(lapply(inputdata, nrow))
-    # sel <- which(!is.na(rast_in[j,,1]) & ll>min_sun*window_size**2)
-    # inputdata <- inputdata[sel]
-    # inputdata <- do.call(what = rbind, args = inputdata)
-
     # compute diversity
     if (!is.null(inputdata)){
       if (length(inputdata)>0){
@@ -120,29 +111,16 @@ get_raster_diversity_mw <- function(
         SSwindows_per_CPU <- split_chunk(SSchunk, nbCPU)
         # 6- compute diversity metrics
         nb_clusters <- dim(Kmeans_info$Centroids[[1]])[1]
-        alphabetaIdx_CPU <- lapply(X = SSwindows_per_CPU$SSwindow_perCPU,
-                                   FUN = alphabeta_window_list,
-                                   nb_clusters = nb_clusters,
-                                   alpha_metrics = alpha_metrics,
-                                   Hill_order = Hill_order,
-                                   Beta_info = Beta_info,
-                                   beta_metrics = beta_metrics,
-                                   pcelim = pcelim)
-        alphabetaIdx <- unlist(alphabetaIdx_CPU, recursive = FALSE)
+        alphabetaIdx <- lapply(X = SSwindows_per_CPU$SSwindow_perCPU,
+                               # FUN = alphabeta_window_list,
+                               FUN = alphabeta_chunk,
+                               nb_clusters = nb_clusters,
+                               alpha_metrics = alpha_metrics,
+                               Hill_order = Hill_order,
+                               Beta_info = Beta_info,
+                               beta_metrics = beta_metrics,
+                               pcelim = pcelim)
 
-        # 6bis- compute functional metrics
-        if (!is.null(fd_metrics)){
-          inputdata <- cbind(center_reduce(x = inputdata[selected_bands],
-                                           m = Kmeans_info$MinVal,
-                                           sig = Kmeans_info$Range),
-                             'win_ID' = inputdata$win_ID)
-          windows_per_CPU <- split_chunk(inputdata, nbCPU)
-          funct_idx_cpu <- lapply(X = windows_per_CPU$SSwindow_perCPU,
-                                  FUN = functional_window_list,
-                                  fd_metrics = fd_metrics)
-          FunctionalIdx <- unlist(funct_idx_cpu, recursive = FALSE)
-          rm(funct_idx_cpu)
-        }
         # 7- reshape alpha diversity metrics
         IDwindow <- unlist(SSwindows_per_CPU$IDwindow_perCPU)
         for (idx in alpha_metrics){
@@ -155,10 +133,6 @@ get_raster_diversity_mw <- function(
             res_shapeChunk[[idx]][[crit]][j,] <- res_shapeChunk_tmp
           }
         }
-
-        # reshape functional
-
-
 
         # resahape beta
         for (beta in beta_metrics){
@@ -175,91 +149,7 @@ get_raster_diversity_mw <- function(
     }
   }
 
-  # # v2: all lines together
-  # nbWindows <- ncol(rast_in)*nrow(rast_in)
-  # xyminmax <- list()
-  # for (j in 1:nrow(rast_in)){
-  #   for (i in 1:ncol(rast_in)){
-  #     cell <- j+(i-1)*nrow(rast_in)
-  #     xyminmax[[cell]] <- list('xmin' = max(c(1, i-(window_size-1)/2)),
-  #                              'xmax' = min(c(ncol(rast_in), i+(window_size-1)/2)),
-  #                              'ymin' = max(c(1, j-(window_size-1)/2)),
-  #                              'ymax' = min(c(nrow(rast_in), j+(window_size-1)/2)))
-  #   }
-  # }
-  #
-  # rastext <- function(xyminmax, cell, rast_in){
-  #   inputdata <- na.omit(rast_in[xyminmax$ymin:xyminmax$ymax, xyminmax$xmin:xyminmax$xmax, ])
-  #   if (nrow(inputdata)>0)
-  #     inputdata$win_ID <- cell
-  #   return(inputdata)
-  # }
-  # cell <- seq_len(length(xyminmax))
-  # inputdata <- mapply(FUN = rastext,
-  #                     xyminmax = xyminmax,
-  #                     cell = cell,
-  #                     MoreArgs = list(rast_in = rast_in))
-  #
-  # # for (j in 1:nrow(rast_in)){
-  # #   print(100*j/nrow(rast_in))
-  # #   for (i in 1:ncol(rast_in)){
-  # #     cell <- j+(i-1)*nrow(rast_in)
-  # #     xmin <- max(c(1, i-(window_size-1)/2))
-  # #     xmax <- min(c(ncol(rast_in), i+(window_size-1)/2))
-  # #     ymin <- max(c(1, j-(window_size-1)/2))
-  # #     ymax <- min(c(nrow(rast_in), j+(window_size-1)/2))
-  # #     inputdata[[cell]] <- na.omit(rast_in[ymin:ymax, xmin:xmax, ])
-  # #     if (nrow(inputdata[[cell]])>0)
-  # #       inputdata[[cell]]$win_ID <- cell
-  # #   }
-  # # }
-  # ll <- unlist(lapply(inputdata, nrow))
-  # sel <- which(!is.na(terra::values(rast_in[[1]])) & ll>min_sun*window_size**2)
-  # inputdata <- inputdata[sel]
-  # inputdata <- do.call(what = rbind, args = inputdata)
-  #
-  # # compute diversity
-  # if (length(sel)>0){
-  #   SSchunk <- get_spectralSpecies(inputdata = inputdata,
-  #                                  Kmeans_info = Kmeans_info,
-  #                                  selected_bands = selected_bands)
-  #   # 5- split data chunk by window and by nbCPU to ensure parallel computing
-  #   rm(inputdata)
-  #   SSwindows_per_CPU <- split_chunk(SSchunk, nbCPU)
-  #   rm(SSchunk)
-  #   gc()
-  #   # 6- compute diversity metrics
-  #   nb_clusters <- dim(Kmeans_info$Centroids[[1]])[1]
-  #   alphabetaIdx_CPU <- lapply(X = SSwindows_per_CPU$SSwindow_perCPU,
-  #                              FUN = alphabeta_window_list,
-  #                              nb_clusters = nb_clusters,
-  #                              alpha_metrics = alpha_metrics,
-  #                              Beta_info = Beta_info,
-  #                              Hill_order = Hill_order,
-  #                              pcelim = pcelim)
-  #   alphabetaIdx <- unlist(alphabetaIdx_CPU,recursive = F)
-  #   rm(alphabetaIdx_CPU)
-  #   gc()
-  #   # 7- reshape alpha diversity metrics
-  #   IDwindow <- unlist(SSwindows_per_CPU$IDwindow_perCPU)
-  #   for (idx in alpha_metrics){
-  #     for (crit in c('mean', 'sd')){
-  #       elem <- paste0(idx,'_',crit)
-  #       restmp <- unlist(lapply(alphabetaIdx,'[[',elem))
-  #       if (length(IDwindow)>0)
-  #         res_shapeChunk[[idx]][[crit]][IDwindow] <- restmp
-  #     }
-  #   }
-  #   PCoA_BC <- matrix(data = NA, nrow = nbWindows, ncol = dimPCO)
-  #   if (!is.null(Beta_info) & !is.null(IDwindow) & length(IDwindow)>0) {
-  #     PCoA_BC0 <- do.call(rbind,lapply(alphabetaIdx,'[[','PCoA_BC'))
-  #     PCoA_BC[IDwindow,] <- PCoA_BC0
-  #   }
-  #   for (i in 1:dimPCO)
-  #     PCoA_raster[[i]] <- matrix(data = PCoA_BC[,i],
-  #                                nrow = nrow(rast_in),
-  #                                ncol = ncol(rast_in))
-  # }
+
   rm(alphabetaIdx)
   gc()
   ab_div_metrics <- list('richness' = res_shapeChunk$richness,