+++ R/kmeans_progressr.R
@@ -6,13 +6,14 @@
 #' @param nstart numeric.
 #' @param algorithm character. algorithm used in the kmeans clustering
 #' @param p function.
+#' @param ... list. additional parameters
 #
 #' @return results of kmeans
 #' @importFrom stats kmeans
 #' @export
 
 kmeans_progressr <- function(x, centers, iter.max, nstart,
-                             algorithm = "Hartigan-Wong", p = NULL){
+                             algorithm = "Hartigan-Wong", p = NULL, ...){
   res <- kmeans(x = x, centers = centers, iter.max = iter.max, nstart = nstart,
                 algorithm = algorithm)
   if (!is.null(p)) p()